+++ examples/functions/inverse_cidr/function.tf
@@ -1,5 +1,0 @@
-# Copyright (c) HashiCorp, Inc.
-
-output "inverse_cidr" {
-  value = provider::iactools::inverse_cidr("192.168.0.0/16", "192.168.1.0/24")
-}