+++ examples/provider/provider.tf
@@ -1,4 +1,6 @@
-# Copyright (c) HashiCorp, Inc.
+# Copyright (c) LederWorks
+# SPDX-FileCopyrightText: The terraform-provider-iactools Authors
+# SPDX-License-Identifier: MPL-2.0
 
 provider "iactools" {
   # requires no configuration