+++ terratest/functions/inverse_cidr/outputs.tf
@@ -2,4 +2,6 @@
 # SPDX-FileCopyrightText: The terraform-provider-iactools Authors
 # SPDX-License-Identifier: MPL-2.0
 
-* @lederworks/iactools
+output "inverse_cidr" {
+  value = provider::iactools::inverse_cidr(var.parent_cidr, var.child_cidr)
+}