+++ terratest/functions/reverse_dns/outputs.tf
@@ -2,4 +2,6 @@
 # SPDX-FileCopyrightText: The terraform-provider-iactools Authors
 # SPDX-License-Identifier: MPL-2.0
 
-* @lederworks/iactools
+output "reverse_dns" {
+  value = provider::iactools::reverse_dns(var.ip_address)
+}