+++ terratest/functions/reverse_dns/variables.tf
@@ -2,4 +2,6 @@
 # SPDX-FileCopyrightText: The terraform-provider-iactools Authors
 # SPDX-License-Identifier: MPL-2.0
 
-* @lederworks/iactools
+variable "ip_address" {
+  type = string
+}