+++ CHANGELOG.md
@@ -1,3 +1,30 @@
+<!--
+Copyright (c) LederWorks
+SPDX-FileCopyrightText: The terraform-provider-iactools Authors
+SPDX-License-Identifier: MPL-2.0
+-->
+
+## 0.2.0 (Released)
+
+FEATURES:
+- Added CodeQL workflow
+- Added Dependabot automerge workflow
+- Added REUSE workflow
+- Added Verify workflow
+- Added GitVersion workflow
+- Added reverse_dns function
+
+ENHANCEMENTS:
+- Added SUPPORT.md
+- Added SECURITY.md
+- Updated README.md
+- Updated CODE_OF_CONDUCT.md
+- Improved documentation
+- Improved testing
+
+BUG FIXES:
+- The inverse_cidr function now calculates CIDR ranges with a Trie Graph
+
 ## 0.1.0 (Unreleased)
 
 FEATURES: